requirements.txt
streamlit==1.32.0
google-genai